feat: Conexão com o MySQl, atualização nos atributos e validações nos métodos POST/DELETE

+++ projetoPI/src/main/java/sptech/school/projetoPI/services/EmployeeService.java
@@ -5,7 +5,9 @@ import org.springframework.stereotype.Service;
 import sptech.school.projetoPI.entities.Employee;
 import sptech.school.projetoPI.exceptions.EntityConflictException;
 import sptech.school.projetoPI.exceptions.EntityNotFoundException;
+import sptech.school.projetoPI.exceptions.ForeignKeyConstraintException;
 import sptech.school.projetoPI.repositories.EmployeeRepository;
+import sptech.school.projetoPI.repositories.ScheduleRepository;
 
 import java.util.List;
 
@@ -13,15 +15,17 @@ import java.util.List;
 public class EmployeeService {
 
     private final EmployeeRepository repository;
+    private final ScheduleRepository scheduleRepository;
 
-    public EmployeeService(EmployeeRepository repository) {
+    public EmployeeService(EmployeeRepository repository, ScheduleRepository scheduleRepository) {
         this.repository = repository;
+        this.scheduleRepository = scheduleRepository;
     }
 
     public Employee signEmployee(Employee employee) {
         if (repository.existsByCpf(employee.getCpf()) || repository.existsByEmail(employee.getEmail()) || repository.existsByPhone(employee.getPhone())) {
             throw new EntityConflictException(
-                    "Este cpf, email ou telefone já está cadastrado no banco."
+                    "Este CPF, email ou telefone já está cadastrado no banco."
             );
         }
 
@@ -45,16 +49,20 @@ public class EmployeeService {
         // Não sei como fazer esse.
     }
 
-    // Não sei se isso tá certo..
     public ResponseEntity<Void> deleteEmployeeById(Integer id) {
         if (!repository.existsById(id)) {
-            throw new EntityConflictException(
-                    "O funcionário de id %d não foi encontrado.".formatted(id)
+            throw new EntityNotFoundException(
+                    "O funcionário de ID %d não foi encontrado".formatted(id)
+            );
+        }
+
+        if(scheduleRepository.existsByEmployeeId(id)) {
+            throw new ForeignKeyConstraintException(
+                    "O funcionário de ID %d não pode ser apagado porque está relacionado com um ou vários agendamentos".formatted(id)
             );
         }
 
         repository.deleteById(id);
         return ResponseEntity.status(204).build();
     }
-
 }
